+++ src/components/hexViewer/Index.vue
@@ -26,6 +26,30 @@
         <q-btn color="white" flat dense :icon="typeOfHexView === 'hex' ? 'mdi-matrix' : 'mdi-format-text'" @click="typeOfHexView = typeOfHexView === 'hex' ? 'text' : 'hex'">
           <q-tooltip>Change view mode (hex/text)</q-tooltip>
         </q-btn>
+        <q-btn color="white" flat dense icon="mdi-export-variant">
+          <q-tooltip>Export</q-tooltip>
+          <q-menu>
+            <q-list style="min-width: 100px" class="bg-grey-8 text-white">
+              <div class="q-pa-sm">
+                <div style="font-size: .8rem">Copy selected as</div>
+                <div>
+                  <q-btn v-close-popup :disable="!selectedMessages.length" dense flat label="seen" @click="copySelected()" />
+                  <q-btn v-close-popup :disable="!selectedMessages.length" dense flat label="hex" @click="copySelected('hex')" />
+                  <q-btn v-close-popup :disable="!selectedMessages.length" dense flat label="raw" @click="copySelected('text')" />
+                </div>
+              </div>
+              <q-separator />
+              <div class="q-pa-sm">
+                <div style="font-size: .8rem">Export selected as</div>
+                <div>
+                  <q-btn v-close-popup :disable="!selectedMessages.length" dense flat label="seen" @click="exportSelected()" />
+                  <q-btn v-close-popup :disable="!selectedMessages.length" dense flat label="hex" @click="exportSelected('hex')" />
+                  <q-btn v-close-popup :disable="!selectedMessages.length" dense flat label="raw" @click="exportSelected('text')" />
+                </div>
+              </div>
+            </q-list>
+          </q-menu>
+        </q-btn>
       </q-toolbar>
       <hex-viewer
         v-if="activeConnection"
@@ -54,6 +78,8 @@ import Messages from '../../components/hexViewer/Messages'
 import MessagePreviewItem from '../../components/hexViewer/MessagePreviewItem'
 import HexViewer from '../HexViewer'
 import { mapState } from 'vuex'
+import { date } from 'quasar'
+import hexProcessing from '../../mixins/hexProcessing'
 export default {
   props: ['active', 'isVisibleToolbar', 'config'],
   data () {
@@ -80,7 +106,44 @@ export default {
       return false
     }
   },
-  methods: {},
+  methods: {
+    getHeader (packet) {
+      const eventsDesc = {
+        0: 'Data recieved',
+        1: 'Connect',
+        2: 'Disconnect'
+      }
+      const info = packet['proxy.source'] === 0 ? 'incoming' : `target ${packet['proxy.source']}`
+      let header = ''
+      header += eventsDesc[packet['proxy.event']]
+      header += `[${info}]`
+      header += '('
+      header += date.formatDate(packet.timestamp * 1000, 'DD/MM/YYYY HH:mm:ss')
+      if (packet['proxy.payload.hex']) {
+        header += ` size: ${packet['proxy.payload.size']}B`
+      }
+      header += ')'
+      return header
+    },
+    getContentBySelected (type) {
+      return this.selectedMessages.reduce((content, packet) => {
+        content += this.getHeader(packet)
+        if (packet['proxy.payload.hex']) {
+          content += this.getContent(type, packet['proxy.payload.hex'], this.typeOfHexView)
+        }
+        content += '\r\n'
+        return content
+      }, '')
+    },
+    copySelected (type = 'view') {
+      const content = this.getContentBySelected(type)
+      this.copy(content, type)
+    },
+    exportSelected (type = 'view') {
+      const content = this.getContentBySelected(type)
+      this.exportData(content, type)
+    }
+  },
   watch: {
     active () {
       this.activeConnection = null
@@ -88,6 +151,7 @@ export default {
       this.connectionPreview = null
     }
   },
+  mixins: [hexProcessing],
   components: { Messages, MessagePreviewItem, HexViewer }
 }
 </script>